refactor(build): trim header dependencies and add IWYU
