+++ .github/workflows/ci.yml
@@ -86,3 +86,11 @@ jobs:
         run: cmake -S . -B build
       - name: Docs
         run: ctest --test-dir build -L Docs --output-on-failure
+  iwyu:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v3
+      - name: Install include-what-you-use
+        run: sudo apt-get update && sudo apt-get install -y include-what-you-use
+      - name: Run IWYU checks
+        run: scripts/run-iwyu.sh
