+++ src/il/io/InstrParser.cpp
@@ -6,11 +6,14 @@
 
 #include "il/io/InstrParser.hpp"
 
-#include "il/io/ParserUtil.hpp"
-#include "il/io/TypeParser.hpp"
+#include "il/core/BasicBlock.hpp"
+#include "il/core/Function.hpp"
+#include "il/core/Instr.hpp"
 #include "il/core/Opcode.hpp"
 #include "il/core/OpcodeInfo.hpp"
 #include "il/core/Value.hpp"
+#include "il/io/ParserUtil.hpp"
+#include "il/io/TypeParser.hpp"
 
 #include <cctype>
 #include <exception>
