+++ src/vm/OpHandlers.cpp
@@ -6,6 +6,9 @@
 
 #include "vm/OpHandlers.hpp"
 
+#include "il/core/BasicBlock.hpp"
+#include "il/core/Function.hpp"
+#include "il/core/Instr.hpp"
 #include "il/core/Opcode.hpp"
 #include "il/core/OpcodeInfo.hpp"
 #include "vm/RuntimeBridge.hpp"
