+++ src/vm/VM.cpp
@@ -5,6 +5,8 @@
 // Links: docs/il-spec.md
 
 #include "vm/VM.hpp"
+#include "il/core/BasicBlock.hpp"
+#include "il/core/Function.hpp"
 #include "il/core/Instr.hpp"
 #include "il/core/Opcode.hpp"
 #include "vm/RuntimeBridge.hpp"
