+++ tests/unit/test_il_parse_extern_missing_arrow.cpp
@@ -4,6 +4,7 @@
 // Ownership/Lifetime: Test constructs modules and buffers locally.
 // Links: docs/il-spec.md
 
+#include "il/core/Module.hpp"
 #include "il/io/Parser.hpp"
 #include <cassert>
 #include <sstream>
