+++ tui/include/tui/ui/widget.hpp
@@ -4,8 +4,12 @@
 // @ownership Derived classes own their state; Widget stores layout rect.
 #pragma once
 
-#include "tui/render/screen.hpp"
-#include "tui/term/input.hpp"
+#include "tui/term/key_event.hpp"
+
+namespace viper::tui::render
+{
+class ScreenBuffer;
+} // namespace viper::tui::render
 
 namespace viper::tui::ui
 {
