+++ tui/tests/test_widgets_basic.cpp
@@ -6,6 +6,7 @@
 #include "tui/render/renderer.hpp"
 #include "tui/render/screen.hpp"
 #include "tui/style/theme.hpp"
+#include "tui/term/key_event.hpp"
 #include "tui/term/term_io.hpp"
 #include "tui/widgets/button.hpp"
 #include "tui/widgets/label.hpp"
